+++ classes/annona/token/TokenGenerateAtype.class.php
@@ -1,13 +1,15 @@
 <?php
 namespace Flex\Annona\Token;
 
-use Flex\Annona\Cipher\Encrypt;
-use Flex\Annona\Cipher\Decrypt;
+use Flex\Annona\Cipher\CipherGeneric;
+use Flex\Annona\Cipher\HashEncoder;
+use Flex\Annona\Cipher\Base64UrlEncoder;
 use Flex\Annona\Token\TokenAbstract;
+
 
 class TokenGenerateAtype extends TokenAbstract
 {
-    public const __version = '1.1';
+    public const __version = '1.2';
     public string $value = '';
 
     public function __construct(string|null $generate_string, int $length=50){
@@ -18,20 +20,20 @@ class TokenGenerateAtype extends TokenAbstract
     public function generateHashKey(string $encrypt_type ='sha512') : TokenGenerateAtype
     {
         $this->value = match ($encrypt_type) {
-            'sha256','sha512' => (new Encrypt($this->value))->_hash($encrypt_type)
+            'sha256','sha512' => (new CipherGeneric(new HashEncoder($this->value)))->hash($encrypt_type)
         };
     return $this;
     }
 
     # @abstract 토큰생성 : _base64_urlencode
     public function generateToken(string $hash) : TokenGenerateAtype {
-        $this->value = (new Encrypt(sprintf("%s%s",$hash,$this->value)))->_base64_urlencode();
+        $this->value = (new CipherGeneric(new Base64UrlEncoder()))->encode(sprintf("%s%s",$hash,$this->value));
     return $this;
     }
 
     # @abstract 토큰생성 : _base64_urlencode
     public function decodeToken(string $token) : TokenGenerateAtype {
-        $this->value = (new Decrypt($token))->_base64_urldecode();
+        $this->value = (new CipherGeneric(new Base64UrlEncoder()))->decode($token);
     return $this;
     }