+++ classes/annona/token/TokenGenerateBtype.class.php
@@ -1,12 +1,13 @@
 <?php
 namespace Flex\Annona\Token;
 
-use Flex\Annona\Cipher\Encrypt;
+use Flex\Annona\Cipher\CipherGeneric;
+use Flex\Annona\Cipher\HashEncoder;
 use Flex\Annona\Token\TokenAbstract;
 
 class TokenGenerateBtype extends TokenAbstract
 {
-    public const __version = '1.1';
+    public const __version = '1.2';
     private string $value = '';
 
     # 랜덤 : AE68A9MPVZ
@@ -14,11 +15,11 @@ class TokenGenerateBtype extends TokenAbstract
         $this->value = $generate_string ?? parent::generateString($length);
     }
 
-    # @abstract 해시키 : sha256
+    # @abstract 해시키 : _md5_base64
     # 5jF4rq3N9V3RLHEBW2RKg
     public function generateHashKey() : TokenGenerateBtype
     {
-        $temp_usertoken = (new Encrypt($this->value))->_hash('sha256');
+        $temp_usertoken = (new CipherGeneric(new HashEncoder($this->value)))->hash('sha256');
         $this->value = parent::cleanEtcWords($temp_usertoken);
     return $this;
     }
@@ -26,7 +27,7 @@ class TokenGenerateBtype extends TokenAbstract
     # @abstract 토큰생성 : SHA512
     # 6e262cc52963a523985f368d1f141e6df34125f1dc03fe28fc9abae8db1f185c3b3f0a81f2e271853f9be4c21a0f35c0cdbb5c9d8486168e14dd60a4337df88f
     public function generateToken(string $hash='') : TokenGenerateBtype {
-        $this->value = (new Encrypt($this->value))->_hash('sha512');
+        $this->value = (new CipherGeneric(new HashEncoder($this->value)))->hash('sha512');
     return $this;
     }